Inline traverseAllChildren into ReactChildren
